Cleaned up variable names and switched to modifying Doc/conf.py directly
