+++ Doc/conf.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath('includes'))
 # ---------------------
 
 extensions = ['sphinx.ext.coverage', 'sphinx.ext.doctest',
-              'pyspecific', 'c_annotations', 'escape4chm', 'ucdversion']
+              'pyspecific', 'c_annotations', 'escape4chm', 'unidata_version']
 
 
 doctest_global_setup = '''
@@ -227,6 +227,6 @@ linkcheck_ignore = [r'https://bugs.python.org/(issue)?\d+',
 # Relative filename of the reference count data file.
 refcount_file = 'data/refcounts.dat'
 
-# Import UCD_VERSION for ucdversion extension.
-with open('includes/UCD_VERSION.txt', 'r') as f:
-    UCD_VERSION = f.read().strip()
+# Unicode data version for unidata_version extension.
+# Managed by Tools/unicode/makeunicodedata.py
+UNIDATA_VERSION = "12.1.0"
