--- Doc/includes/UCD_VERSION.txt
@@ -1,1 +1,0 @@
-12.1.0
